requirements.txt
python-dotenv==0.21.0
uvicorn==0.20.0
mysql-connector-python==8.0.31
